[android] updated bindings for v4.3.0

+++ BlinkID/src/android/java/com/phonegap/plugins/microblink/recognizers/serialization/CroatiaCombinedRecognizerSerialization.java
@@ -32,7 +32,7 @@ public final class CroatiaCombinedRecognizerSerialization implements RecognizerS
             jsonResult.put("dateOfExpiryPermanent", result.isDateOfExpiryPermanent());
             jsonResult.put("dateOfIssue", SerializationUtils.serializeDate(result.getDateOfIssue()));
             jsonResult.put("digitalSignature", SerializationUtils.encodeByteArrayToBase64(result.getDigitalSignature()));
-            jsonResult.put("digitalSignatureVersion", result.getDigitalSignatureVersion());
+            jsonResult.put("digitalSignatureVersion", (int)result.getDigitalSignatureVersion());
             jsonResult.put("documentBilingual", result.isDocumentBilingual());
             jsonResult.put("documentDataMatch", result.isDocumentDataMatch());
             jsonResult.put("faceImage", SerializationUtils.encodeImageBase64(result.getFaceImage()));
